[MRELEASE-232] Fix NPE
git-svn-id: https://svn.apache.org/repos/asf/maven/release/trunk@539088 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-release-manager/src/main/java/org/apache/maven/shared/release/phase/AbstractBackupPomsPhase.java
@@ -35,14 +35,21 @@ public abstract class AbstractBackupPomsPhase
     {
         File pomFile = project.getFile();
 
-        return new File( pomFile.getAbsolutePath() + backupPrefix );
+        if ( pomFile != null )
+        {
+            return new File( pomFile.getAbsolutePath() + backupPrefix );
+        }
+        else
+        {
+            return null;
+        }
     }
 
     protected void deletePomBackup( MavenProject project )
     {
         File pomBackup = getPomBackup( project );
 
-        if ( pomBackup.exists() )
+        if ( pomBackup != null && pomBackup.exists() )
         {
             pomBackup.delete();
         }
